moved selectors to separate file
Updated connected components
