+++ client/src/components/user/UserCourses.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Moment from 'react-moment';
-import { selectUserCourseNames } from '../../reducers';
+import { selectUserCourseNames } from '../../reducers/selectors';
 
 class UserCourses extends Component {
   renderCourses(courses) {
